Reorg: mv new types and system requests to sep files
