+++ lua/stache/ask.lua
@@ -1,0 +1,30 @@
+local function ask_cr(cmd)
+    local res = {}
+    local function collect(_, data, name)
+        res[name] = data
+    end
+    local j = vim.fn.jobstart(cmd, {
+        on_stdout = collect,
+        on_stderr = collect,
+        stdout_buffered = true,
+        stderr_buffered = true,
+    })
+    vim.fn.jobwait({ j })
+    coroutine.yield(res)
+end
+
+local function ask_rg(args)
+    local command = { "rg" }
+    for _, arg in ipairs(args) do
+        table.insert(command, arg)
+    end
+    local co = coroutine.create(ask_cr)
+    local _, ans = coroutine.resume(co, command)
+    assert(coroutine.resume(co))
+    return ans
+end
+
+return {
+    ask_cr = ask_cr,
+    ask_rg = ask_rg,
+}
